add proper types for visual-helpers, add export point for  visual-helpers,
