+++ src/visual-helpers/styleguide-device-frame/styleguide-device-frame.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import StyleguideCell from './styleguide-cell';
+import StyleguideCell from '../styleguide-cell/styleguide-cell';
 
 const SIZES = {
     SMALL: 'SMALL',
