+++ src/visual-helpers/styleguide-device-range/styleguide-device-range.d.ts
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export type StyleguideDeviceRangeProps = {
+    /**
+     * The elements to be inserted in the content block
+     */
+    children?: React.ReactNode;
+};
+
+declare const StyleguideDeviceRange: React.FunctionComponent<StyleguideDeviceRangeProps>;
+
+export default StyleguideDeviceRange;
