+++ src/visual-helpers/styleguide-static.js
